requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3
